+++ src/store/route/selectors.ts
@@ -1,11 +1,10 @@
 import { createSelector } from 'reselect';
 
 import { LoadableValue, RouteState, StoreState } from '../types';
-import { routeDataQuery, routeParams } from '../../state/route';
 
 const routeState = (state: StoreState): RouteState => state.route;
 
-const routeValueState = createSelector(routeState, (state) => state.route);
+export const routeSelector = createSelector(routeState, (state) => state.route);
 
 const isInitial = (item: LoadableValue): boolean => item.state === 'initial';
 const isLoading = (item: LoadableValue): boolean => item.state === 'loading';
@@ -14,17 +13,18 @@ const getError = (item: LoadableValue): Error | undefined => (item.state === 'er
 const hasValue = (item: LoadableValue): boolean => item.state === 'hasValue';
 const getValue = <T>(item: LoadableValue<T>): T | undefined => (item.state === 'hasValue' ? item.data : undefined);
 
-const isRouteInitialSelector = createSelector(routeValueState, isInitial);
-export const isRouteLoadingSelector = createSelector(routeValueState, isLoading);
-const hasRouteErrorSelector = createSelector(routeValueState, hasError);
-const routeErrorSelector = createSelector(routeValueState, getError);
-const hasRouteValueSelector = createSelector(routeValueState, hasValue);
-const getRouteValueSelector = createSelector(routeValueState, getValue);
+const isRouteInitialSelector = createSelector(routeSelector, isInitial);
+export const isRouteLoadingSelector = createSelector(routeSelector, isLoading);
+const hasRouteErrorSelector = createSelector(routeSelector, hasError);
+const routeErrorSelector = createSelector(routeSelector, getError);
+const hasRouteValueSelector = createSelector(routeSelector, hasValue);
+const getRouteValueSelector = createSelector(routeSelector, getValue);
 
 const distanceSelector = createSelector(routeState, (state) => state.distance);
 const routeTypeSelector = createSelector(routeState, (state) => state.routeType);
 const locationSelector = createSelector(routeState, (state) => state.location);
 const randomSeedSelector = createSelector(routeState, (state) => state.randomSeed);
+export const flippedSelector = createSelector(routeState, (state) => state.flipped);
 
 export const routeParametersSelector = createSelector(
   [distanceSelector, routeTypeSelector, locationSelector, randomSeedSelector],
